Fixed some bugs
security issue(edit video,delete video)=> only owner

delete issue => delete from the db.

+++ src/controllers/videoController.js
@@ -20,11 +20,18 @@ export const watch = async (req, res) => {
   return res.render("videos/watch", { pageTitle: video.title, video });
 };
 export const getEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
   const video = await Video.findById(id);
 
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  console.log(video.owner);
+  if (video.owner != _id) {
+    return res.status(403).redirect("/");
   }
   return res.render("videos/edit", {
     pageTitle: `Edit ${video.title} `,
@@ -32,11 +39,18 @@ export const getEdit = async (req, res) => {
   });
 };
 export const postEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
-  const video = await Video.exists({ _id: id });
+  const video = await Video.findById(id);
   if (!video) {
     return res.render("404", { pageTitle: "Video not found" });
+  }
+
+  if (video.owner != _id) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -77,8 +91,21 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  if (video.owner != _id) {
+    return res.status(403).render("/");
+  }
   await Video.findByIdAndDelete(id);
+  user.videos.splice(user.videos.indexOf(id), 1);
+  user.save();
   return res.redirect("/");
 };
